# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyTeacher Mobile API
  version: 1.0.0
  description: |
    MyTeacherモバイルアプリ用バックエンドAPI
    
    ## 認証方式
    AWS Cognito JWT認証を使用します。
    すべてのエンドポイントは以下のヘッダーが必要です：
    ```
    Authorization: Bearer {cognito_jwt_token}
    ```
    
    ## ベースURL
    - 本番環境: `https://my-teacher-app.com/api/v1`
    - ローカル開発: `http://localhost:8080/api/v1`
    
    ## レスポンス形式
    すべてのレスポンスは以下の形式を持ちます：
    
    **成功時**:
    ```json
    {
      "success": true,
      "message": "処理が完了しました",
      "data": { ... }
    }
    ```
    
    **エラー時**:
    ```json
    {
      "success": false,
      "message": "エラーメッセージ",
      "errors": {
        "field_name": ["エラー詳細"]
      }
    }
    ```
    
  contact:
    name: MyTeacher Development Team
    email: famicoapp@gmail.com
  license:
    name: Proprietary
    url: https://my-teacher-app.com/license
servers:
  - url: https://my-teacher-app.com/api/v1
    description: 本番環境
  - url: http://localhost:8080/api/v1
    description: ローカル開発環境
...
security:
  - CognitoAuth: []
  - SanctumAuth: []
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWTトークン（Web版用）
    SanctumAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: Laravel Sanctumトークン（モバイル版用）
...
  responses:
    UnauthorizedError:
      description: 認証エラー - JWT/Sanctumトークンが無効または期限切れ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "認証が必要です"
...
paths:
...
  /auth/login:
    post:
      summary: ログイン（モバイル用）
      description: |
        usernameとpasswordでログインし、Sanctumトークンを取得します。
        
        **注意**: Web版はCognito JWT認証、モバイル版はSanctum認証を使用します。
      tags: [Authentication]
      security: []  # 認証不要
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  description: ユーザー名（メールアドレスではない）
                  example: "taro_tanaka"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "password123"
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Laravel Sanctumトークン（有効期限30日）
                    example: "1|abcdefghijklmnopqrstuvwxyz123456"
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "田中太郎"
                      email:
                        type: string
                        format: email
                        example: "taro@example.com"
                      username:
                        type: string
                        example: "taro_tanaka"
                      avatar_url:
                        type: string
                        nullable: true
                        example: "https://s3.amazonaws.com/avatars/123.png"
                      created_at:
                        type: string
                        format: date-time
        '401':
          description: 認証失敗 - usernameまたはpasswordが間違っている
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "認証に失敗しました"
                  errors:
                    type: object
                    properties:
                      username:
                        type: array
                        items:
                          type: string
                        example: ["ユーザー名またはパスワードが間違っています"]
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      summary: ログアウト（モバイル用）
      description: 現在のSanctumトークンを無効化します。
      tags: [Authentication]
      security:
        - SanctumAuth: []
      responses:
        '200':
          description: ログアウト成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "ログアウトしました"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================================
  # Tasks API（14エンドポイント）
  # ============================================================
  /tasks:
    get:
      summary: タスク一覧取得
      description: ユーザーの全タスクを取得します。
      tags: [Tasks]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: タスク作成
      description: 新しいタスクを作成します。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  example: "宿題をする"
                description:
                  type: string
                  example: "数学の問題集p.20-25"
                due_date:
                  type: string
                  format: date
                  example: "2025-12-10"
                priority:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 3
                tags:
                  type: array
                  items:
                    type: integer
                  example: [1, 2]
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Task'
        '422':
          $ref: '#/components/responses/ValidationError'
  /tasks/paginated:
    get:
      summary: タスク一覧取得（ページネーション付き）
      description: 無限スクロール用のページネーション付きタスク一覧を取得します。
      tags: [Tasks]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /tasks/{task}:
    put:
      summary: タスク更新
      description: 既存のタスクを更新します。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                due_date:
                  type: string
                  format: date
                priority:
                  type: integer
                tags:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: タスク削除
      description: タスクを削除します。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tasks/{task}/toggle:
    patch:
      summary: タスク完了状態切り替え
      description: タスクの完了/未完了を切り替えます。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tasks/{task}/approve:
    post:
      summary: タスク承認
      description: グループタスクを承認します。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 承認成功
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /tasks/{task}/reject:
    post:
      summary: タスク却下
      description: グループタスクを却下します。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "もう一度確認してください"
      responses:
        '200':
          description: 却下成功
  /tasks/{task}/request-approval:
    post:
      summary: タスク完了申請
      description: タスクの完了を申請します（承認が必要な場合）。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 申請成功
  /tasks/{task}/images:
    post:
      summary: タスク画像アップロード
      description: タスクに画像を添付します。
      tags: [Tasks]
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: アップロード成功
  /task-images/{image}:
    delete:
      summary: タスク画像削除
      description: タスクに添付された画像を削除します。
      tags: [Tasks]
      parameters:
        - name: image
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
  /tasks/bulk-complete:
    patch:
      summary: タスク一括完了
      description: 複数のタスクを一括で完了/未完了にします。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_ids, completed]
              properties:
                task_ids:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
                completed:
                  type: boolean
                  example: true
      responses:
        '200':
          description: 成功
  /tasks/search:
    post:
      summary: タスク検索
      description: キーワードでタスクを検索します。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword:
                  type: string
                  example: "宿題"
                tags:
                  type: array
                  items:
                    type: integer
                completed:
                  type: boolean
      responses:
        '200':
          description: 成功
...
  /groups:
    patch:
      summary: グループ情報更新
      description: グループ名を更新します。
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "田中家"
      responses:
        '200':
          description: 更新成功
  /groups/members:
    post:
      summary: メンバー追加
      description: 招待コードでグループにメンバーを追加します。
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "member@example.com"
      responses:
        '201':
          description: 追加成功
  /groups/members/{member}/permission:
    patch:
      summary: メンバー権限更新
      description: メンバーのタスク割り当て権限を更新します。
      tags: [Groups]
      parameters:
        - name: member
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [can_assign_tasks]
              properties:
                can_assign_tasks:
                  type: boolean
                  example: true
      responses:
        '200':
          description: 更新成功
  /groups/members/{member}/theme:
    patch:
      summary: メンバーテーマ切り替え
      description: メンバーのテーマを変更します。
      tags: [Groups]
      parameters:
        - name: member
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 変更成功
  /groups/transfer/{newMaster}:
    post:
      summary: グループマスター譲渡
      description: グループのマスター権限を他のメンバーに譲渡します。
      tags: [Groups]
      parameters:
        - name: newMaster
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 譲渡成功
  /groups/members/{member}:
    delete:
      summary: メンバー削除
      description: グループからメンバーを削除します。
      tags: [Groups]
      parameters:
        - name: member
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功

  # ============================================================
  # Profile API（5エンドポイント）
  # ============================================================
...
  /profile:
    patch:
      summary: プロフィール更新
      description: ユーザーのプロフィール情報を更新します。
      tags: [Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "田中太郎"
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 更新成功

    delete:
      summary: アカウント削除
      description: ユーザーアカウントを削除します。
      tags: [Profile]
      responses:
        '200':
          description: 削除成功
  /profile/timezone:
    get:
      summary: タイムゾーン取得
      description: ユーザーのタイムゾーン設定を取得します。
      tags: [Profile]
      responses:
        '200':
          description: 成功

    put:
      summary: タイムゾーン更新
      description: ユーザーのタイムゾーンを更新します。
      tags: [Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timezone]
              properties:
                timezone:
                  type: string
                  example: "Asia/Tokyo"
      responses:
        '200':
          description: 更新成功

  # ============================================================
  # Tags API（4エンドポイント）
  # ============================================================
  /tags:
    get:
      summary: タグ一覧取得
      description: ユーザーの全タグを取得します。
      tags: [Tags]
      responses:
        '200':
          description: 成功

    post:
      summary: タグ作成
      description: 新しいタグを作成します。
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, color]
              properties:
                name:
                  type: string
                  example: "宿題"
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                  example: "#3B82F6"
      responses:
        '201':
          description: 作成成功
  /tags/{id}:
    put:
      summary: タグ更新
      description: タグの名前または色を更新します。
      tags: [Tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '200':
          description: 更新成功

    delete:
      summary: タグ削除
      description: タグを削除します。
      tags: [Tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功

  # ============================================================
  # Avatars API（7エンドポイント）
  # ============================================================
  /avatar:
    get:
      summary: アバター情報取得
      description: ユーザーのアバター情報を取得します。
      tags: [Avatars]
      responses:
        '200':
          description: 成功

    post:
      summary: アバター作成
      description: 新しいアバターを作成します（AI生成）。
      tags: [Avatars]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "さくら先生"
                is_transparent:
                  type: boolean
                  example: false
      responses:
        '201':
          description: 作成成功（非同期処理）

    put:
      summary: アバター更新
      description: アバターの設定を更新します。
      tags: [Avatars]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                is_transparent:
                  type: boolean
      responses:
        '200':
          description: 更新成功

    delete:
      summary: アバター削除
      description: アバターを削除します。
      tags: [Avatars]
      responses:
        '200':
          description: 削除成功
...
  /notifications:
    get:
      summary: 通知一覧取得
      description: ユーザーの全通知を取得します。
      tags: [Notifications]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 成功
  /notifications/{id}:
    get:
      summary: 通知詳細取得
      description: 特定の通知の詳細を取得します。
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
...